materialised-view-creation/requirements.txt
python-dotenv
rdflib